Gui lite update 02.5
+GuiDebugCommon
+GuiDebugPhys
GuiText no more auto-text-bound-update.
Better GuiMenubar. GuiMenu-Item: slider, divider, bswitch
Slider, Button new little size
- GuiScreenDebug
- GuiScreenInGame
- GuiScreenRoot,
- GuiScreenChatMessages

discover problem, blender models is different to ModelRenderer.MODELS.

+++ src/outskirts/client/gui/debug/GuiScreen3DVertices.java
@@ -93,7 +93,7 @@ public class GuiScreen3DVertices extends GuiScreen {
     private GuiSlider _VertDisplayNumControl = addGui(new GuiSlider())
             .setUserMinMaxValue(0, 200)
             .addValueChangedListener(e -> {
-                e.<GuiSlider>gui().setText("DisplayNum: "+e.<GuiSlider>gui().getCurrentUserValue());
+                e.<GuiSlider>gui().setText("DisplayNum: "+(int)e.<GuiSlider>gui().getCurrentUserValue());
             }).setValue(1f);
 
     public static Vert addVert(String name, Vector3f position, Vector4f color) {
